Enhance README and improve commit organization workflows
